requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
astroid==2.3.3
asyncio==3.4.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
chardet==3.0.4
click==8.1.7
colorama==0.4.6
coverage==7.4.4
dill==0.3.8
disutils==1.4.32.post2
Faker==14.0.0
Flask==2.0.3
Flask-Cors==3.0.8
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==20.1.0
idna==2.8
iniconfig==2.0.0
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.4.3
MarkupSafe==2.1.5
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
mccabe==0.6.1
more-itertools==10.2.0
multidict==6.0.5
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.5
py==1.11.0
pydub==0.25.1
PyJWT==2.1.0
pylint==2.4.4
pytest==8.1.1
pytest-cov==4.1.0
pytest-env==0.6.2
python-dateutil==2.9.0.post0
python-decouple==3.6
pytz==2022.6
redis==5.0.3
requests==2.31.0
rq==1.1.0
setuptools==69.5.1
six==1.16.0
SQLAlchemy==1.4.15
tomlkit==0.12.4
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==2.0.3
wrapt==1.11.2
yarl==1.9.4
